requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
google-genai==1.55.0
pandas-gbq==0.31.1
google-cloud-bigquery==3.38.0
google-cloud-pubsub==2.33.0
google-cloud-aiplatform==1.130.0
python-dotenv==1.2.1